+++ src/main/java/racingcar/dao/GameDao.java
@@ -1,34 +1,12 @@
 package racingcar.dao;
 
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import racingcar.dao.entity.GameEntity;
 
 import java.util.List;
 
-@Repository
-public class GameDao {
-    private final SimpleJdbcInsert insertActor;
-    private final JdbcTemplate jdbcTemplate;
-    public GameDao(final JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.insertActor = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("game")
-                .usingGeneratedKeyColumns("game_id");
-    }
+public interface GameDao {
 
-    public int save(GameEntity gameEntity) {
-        SqlParameterSource sqlParameterSource = new BeanPropertySqlParameterSource(gameEntity);
-        return insertActor.executeAndReturnKey(sqlParameterSource).intValue();
-    }
+    int save(GameEntity gameEntity);
 
-    public List<Integer> findGameIds() {
-        String sql = "SELECT game_id FROM GAME";
-        return jdbcTemplate.query(sql,(resultSet, rowNum)-> {
-            return resultSet.getInt("game_id");
-        });
-    }
+    List<Integer> findGameIds();
 }
